+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # Changelog
 
 
+## [0.21] - Unreleased
+
+- New customizable user option `pgmacs-large-database-threshold`. For a database larger than this
+  value (queried via `pg_database_size`), PGmacs will estimate table row counts using an imprecise
+  method that does not require a full table scan, but will provide invalid results for tables that
+  have not been VACUUMed or ANALYZEd. For sizes below this threshold, a more accurate `SELECT
+  COUNT(*) FROM table_name` query will be used. If set to zero, full table scans will never be
+  issued (this may be a good choice on large production databases).
+
+- Further workarounds for semi-compatible PostgreSQL variants that don't implement all the system
+  tables that we query to obtain metainformation concerning sizes, contraints, comments and so on.
+  PGmacs provides a more spartan display for such databases.
+
+
 ## [0.20] - 2025-01-02
 
 - New variable `pgmacs-row-list-buttons` which offers similar functionality to